fix(weather-data): upgrade to ng20 creates conflict on circular read of route definitions
