+++ src/app/shared/modules/weather-data/models/route-tokens.ts
@@ -1,0 +1,6 @@
+export const ROUTE_TOKENS = {
+    home: 'weather',
+    stations: 'stations',
+    annualStation: 'annual-station',
+    fileImport: 'file-import'
+} as const;
